+++ app/admin/announcement/create/AnnouncementCreateClient.tsx
@@ -1,13 +1,14 @@
+"use client";
 
-'use client';
+import AnnouncementForm from "../_components/AnnouncementForm";
 
-import AnnouncementForm from '../_components/AnnouncementForm';
-
-export function ClientComponent() {
-    return (
-        <div className="p-6">
-            <h1 className="text-2xl font-bold mb-6 dark:text-white">Create New Announcement</h1>
-            <AnnouncementForm />
-        </div>
-    );
+export function AnnouncementCreateClient() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-bold mb-6 dark:text-white">
+        Buat Pengumuman Baru
+      </h1>
+      <AnnouncementForm />
+    </div>
+  );
 }