+++ app/admin/announcement/create/page.tsx
@@ -1,7 +1,7 @@
-import { ensurePermission } from '@/lib/rbac'
-import { ClientComponent } from './AnnouncementCreateClient'
+import { ensurePermission } from "@/lib/rbac";
+import { AnnouncementCreateClient } from "./AnnouncementCreateClient";
 
 export default async function Page() {
-    await ensurePermission('announcement:create')
-    return <ClientComponent />
+  await ensurePermission("announcement:create");
+  return <AnnouncementCreateClient />;
 }